Adding commands to run `axiom-configure` in Fedora.
This adds the commands to install the requirements in Fedora. As
`doctl` is not available a direct download from github was used.
Also, `lsb_release` is not installed by default so the `$OS`
variable might cause a failure. So, in order to use this patch the
`redhat-lsb-release` package must be installed. Another option could
be to parse the content of `/etc/os-release` but I'm not sure if that
works in Debian/Ubuntu and Arch.