[setup][configure] Check for lsb_release binary
Check for the lsb_release binary that might be missing from the
system.
As we can't determine which OS is running I think this should be a
"fatal" error and the setup should stop.
This commit also includes an update to the README as well as a
alternate path to ZSH location.