Update axiom-config (#303)
* git package repeated for Ubuntu and WSL

During `apt-get install` for dependencies the git package is repeated.
No error caused but the duplication has been removed.

* Move the install of `go` lower

Currently on macOS `axiom-configure` will download and extract the Linux
version of Go and install it under `/usr/local`. Then in the macOS `go`
will be downloaded by Homebrew. Moving the block means that the `go`
command will be found on macOS and skip the download.

* Remove archives after they have been extracted

Currently, the configure script does not remove any archives and remain on
disk. This change removes them.

* Upgrade Go from 1.14.4 --> 1.16.3

* Fedora repeat `dnf install` for fzf and git

This commit removes two lines that look to be duplicates for Fedora install.
Additionally, the `dnf install` was missing the `-y` requiring user input.

* Install gowitness on Fedora