axiom-scan enhancements (#658)
* Update axiom-scan to handle spaces in args

If one is passing something like a user-agent string to the gowitness module, axiom-scan must recognize the use of spaces and double quote the argument before appending it to the args string

* Updated axiom-scan to fix gowitness server command

The method of starting the gowitness server was out-of-date. This new command also points to gowitness server to the proper output directory where the DB and screenshots are located.